backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
streamlit
pandas
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
